compiler fixes

+++ barretenberg/cpp/src/barretenberg/proof_system/plookup_tables/fixed_base/fixed_base.cpp
@@ -193,7 +193,7 @@ BasicTable table::generate_basic_fixed_base_table(BasicTableId id, size_t basic_
     const bool is_small_table = (multitable_bits - bits_covered_by_previous_tables_in_multitable) < BITS_PER_TABLE;
     const size_t table_bits =
         is_small_table ? multitable_bits - bits_covered_by_previous_tables_in_multitable : BITS_PER_TABLE;
-    const size_t table_size = 1ULL << table_bits;
+    const size_t table_size = static_cast<size_t>(1ULL << table_bits);
     BasicTable table;
     table.id = id;
     table.table_index = basic_table_index;
